Remove default commit_rows and vector_matrix_product impls from Polynomial trait; remove evaluate and get from Polynomial trait
